Mixer

+++ globals.h
@@ -1,2 +1,2 @@
-extern int global_chunk_size;
+extern size_t global_chunk_size;
 extern double global_frame_rate;
